refactor: PresentationSampleScreen の責務を分割
PresentationSampleScreen に variant 分岐 (opaque / transparent-overlay)、
isInTab による SafeAreaView の edges 制御、本文の組み立て、セクション描画、
ナビゲーションが同居して論理的凝集になっていたため、責務単位で分割する。

variant 軸では opaque 専用の PresentationSampleScreen と transparent-overlay
専用の PresentationSampleOverlay を別ファイルに分離し、variant prop を撤廃。
共通本体を PresentationSampleBody、セクションカードを PresentationSection
としてそれぞれ独立させた。

isInTab については SafeAreaView 自体を呼び出し側 (on-root) に持たせる構造に
切り替え、コンポーネントから「タブ内/ルート」のレイアウトコンテキストの知識を
取り除いた。NativeTabs (expo-router/unstable-native-tabs) は
BottomTabBarHeightContext を伝播させないため context 経由の自動判定は採れず、
配置情報を呼び出し側が宣言する形が最も明快と判断した。

呼び出し側 14 箇所:
- opaque 系 on-root 5 箇所: SafeAreaView edges=["bottom"] でラップ
- opaque 系 in-tab 5 箇所: isInTab prop を削除 (タブ内では SafeAreaView 不要)
- overlay 系 4 箇所: import を PresentationSampleOverlay に差し替え、
  variant prop を削除

lingui の文言・キーに変更はなく、messages.po は参照行の更新のみ。
