+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/form-sheet/in-tab.tsx
@@ -10,7 +10,6 @@ export default function FormSheetSample(): ReactElement {
     <>
       <Stack.Screen.Title>{t`formSheet`}</Stack.Screen.Title>
       <PresentationSampleScreen
-        isInTab
         presentationValue="formSheet"
         heading={<Trans>タブ内 Stack 上のシート表示</Trans>}
         iosBehavior={
