+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/modal/in-tab.tsx
@@ -10,7 +10,6 @@ export default function ModalSample(): ReactElement {
     <>
       <Stack.Screen.Title>{t`modal`}</Stack.Screen.Title>
       <PresentationSampleScreen
-        isInTab
         presentationValue="modal"
         heading={<Trans>タブ内 Stack 上のモーダル表示</Trans>}
         iosBehavior={
